# Add buffer pool and stateful errors