--- a/reader-v1.go
+++ b/reader-v1.go
@@ -18,7 +18,6 @@ import (
 	"errors"
 	"io"
 	"io/ioutil"
-	"sync"
 )
 
 type encReaderV10 struct {
@@ -28,6 +27,7 @@ type encReaderV10 struct {
 	buffer      packageV10
 	offset      int
 	payloadSize int
+	stateErr    error
 }
 
 // encryptReaderV10 returns an io.Reader wrapping the given io.Reader.
@@ -40,12 +40,20 @@ func encryptReaderV10(src io.Reader, config *Config) (*encReaderV10, error) {
 	return &encReaderV10{
 		authEncV10:  ae,
 		src:         src,
-		buffer:      make(packageV10, maxPackageSize),
+		buffer:      packageBufferPool.Get().([]byte)[:maxPackageSize],
 		payloadSize: config.PayloadSize,
 	}, nil
 }
 
+func (r *encReaderV10) recycle() {
+	recyclePackageBufferPool(r.buffer)
+	r.buffer = nil
+}
+
 func (r *encReaderV10) Read(p []byte) (int, error) {
+	if r.stateErr != nil {
+		return 0, r.stateErr
+	}
 	var n int
 	if r.offset > 0 { // write the buffered package to p
 		remaining := r.buffer.Length() - r.offset // remaining encrypted bytes
@@ -61,6 +69,8 @@ func (r *encReaderV10) Read(p []byte) (int, error) {
 	for len(p) >= headerSize+r.payloadSize+tagSize {
 		nn, err := io.ReadFull(r.src, p[headerSize:headerSize+r.payloadSize]) // read plaintext from src
 		if err != nil && err != io.ErrUnexpectedEOF {
+			r.recycle()
+			r.stateErr = err
 			return n, err // return if reading from src fails or reached EOF
 		}
 		r.Seal(p, p[headerSize:headerSize+nn])
@@ -70,6 +80,8 @@ func (r *encReaderV10) Read(p []byte) (int, error) {
 	if len(p) > 0 {
 		nn, err := io.ReadFull(r.src, r.buffer[headerSize:headerSize+r.payloadSize]) // read plaintext from src
 		if err != nil && err != io.ErrUnexpectedEOF {
+			r.stateErr = err
+			r.recycle()
 			return n, err // return if reading from src fails or reached EOF
 		}
 		r.Seal(r.buffer, r.buffer[headerSize:headerSize+nn])
@@ -87,8 +99,9 @@ type decReaderV10 struct {
 	authDecV10
 	src io.Reader
 
-	buffer packageV10
-	offset int
+	buffer   packageV10
+	offset   int
+	stateErr error
 }
 
 // decryptReaderV10 returns an io.Reader wrapping the given io.Reader.
@@ -101,11 +114,19 @@ func decryptReaderV10(src io.Reader, config *Config) (*decReaderV10, error) {
 	return &decReaderV10{
 		authDecV10: ad,
 		src:        src,
-		buffer:     make(packageV10, maxPackageSize),
+		buffer:     packageBufferPool.Get().([]byte)[:maxPackageSize],
 	}, nil
 }
 
+func (r *decReaderV10) recycle() {
+	recyclePackageBufferPool(r.buffer)
+	r.buffer = nil
+}
+
 func (r *decReaderV10) Read(p []byte) (n int, err error) {
+	if r.stateErr != nil {
+		return 0, r.stateErr
+	}
 	if r.offset > 0 { // write the buffered plaintext to p
 		payload := r.buffer.Payload()
 		remaining := len(payload) - r.offset // remaining plaintext bytes
@@ -120,10 +141,14 @@ func (r *decReaderV10) Read(p []byte) (n int, err error) {
 	}
 	for len(p) >= maxPayloadSize {
 		if err = r.readPackage(r.buffer); err != nil {
+			r.stateErr = err
+			r.recycle()
 			return n, err
 		}
 		length := len(r.buffer.Payload())
 		if err = r.Open(p[:length], r.buffer[:r.buffer.Length()]); err != nil { // notice: buffer.Length() may be smaller than len(buffer)
+			r.stateErr = err
+			r.recycle()
 			return n, err // decryption failed
 		}
 		p = p[length:]
@@ -131,10 +156,14 @@ func (r *decReaderV10) Read(p []byte) (n int, err error) {
 	}
 	if len(p) > 0 {
 		if err = r.readPackage(r.buffer); err != nil {
+			r.stateErr = err
+			r.recycle()
 			return n, err
 		}
 		payload := r.buffer.Payload()
 		if err = r.Open(payload, r.buffer[:r.buffer.Length()]); err != nil { // notice: buffer.Length() may be smaller than len(buffer)
+			r.stateErr = err
+			r.recycle()
 			return n, err // decryption failed
 		}
 		if len(payload) < len(p) {
@@ -170,8 +199,7 @@ func (r *decReaderV10) readPackage(dst packageV10) error {
 type decReaderAtV10 struct {
 	src io.ReaderAt
 
-	ad      authDecV10
-	bufPool sync.Pool
+	ad authDecV10
 }
 
 // decryptReaderAtV10 returns an io.ReaderAt wrapping the given io.ReaderAt.
@@ -185,12 +213,7 @@ func decryptReaderAtV10(src io.ReaderAt, config *Config) (*decReaderAtV10, error
 		ad:  ad,
 		src: src,
 	}
-	r.bufPool = sync.Pool{
-		New: func() interface{} {
-			b := make([]byte, maxPackageSize)
-			return &b
-		},
-	}
+
 	return r, nil
 }
 
@@ -204,12 +227,12 @@ func (r *decReaderAtV10) ReadAt(p []byte, offset int64) (n int, err error) {
 		return 0, errUnexpectedSize
 	}
 
-	buffer := r.bufPool.Get().(*[]byte)
-	defer r.bufPool.Put(buffer)
+	buffer := packageBufferPool.Get().([]byte)[:maxPackageSize]
+	defer recyclePackageBufferPool(buffer)
 	decReader := decReaderV10{
 		authDecV10: r.ad,
 		src:        &sectionReader{r.src, t * maxPackageSize},
-		buffer:     packageV10(*buffer),
+		buffer:     packageV10(buffer),
 		offset:     0,
 	}
 	decReader.SeqNum = uint32(t)